requirements.txt
videosdk-agents[deepgram,openai,elevenlabs,silero,turn_detector]
openai>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
